arreglo errores en el modelado de la base de datos y se implemento los cambios en la app

+++ app/src/main/java/daniel/ornelas/tianguisapp/data/model/CompraModel.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class CompraModel(
         @PrimaryKey(autoGenerate = true)
         val idCompra: Long,
-        val cantProductos: Int,
+        val cantProductos: Long,
         val montoTotal: Float,
         val fecha: String
 ) {
